+++ packages/store/src/lib/store.ts
@@ -1,5 +1,7 @@
 import { BehaviorSubject, Observable } from 'rxjs';
 import { Query } from '..';
+import { batchInProgress, batchDone$ } from './batch';
+import { elfHooksRegistry } from './elf-hooks';
 import { addStore, removeStore } from './registry';
 
 export class Store<
@@ -7,6 +9,8 @@ export class Store<
   State = SDef['state']
 > extends BehaviorSubject<State> {
   initialState!: State;
+  state!: State;
+  private batchInProgress = false;
 
   private context: ReducerContext = {
     config: this.getConfig(),
@@ -14,6 +18,7 @@ export class Store<
 
   constructor(private storeDef: SDef) {
     super(storeDef.state);
+    this.state = storeDef.state;
     this.initialState = this.getValue();
     addStore(this);
   }
@@ -33,15 +38,38 @@ export class Store<
   update(...reducers: Array<Reducer<State>>) {
     const currentState = this.getValue();
 
-    const nextState = reducers.reduce((value, reducer) => {
+    let nextState = reducers.reduce((value, reducer) => {
       value = reducer(value, this.context);
 
       return value;
     }, currentState);
 
-    if (nextState !== currentState) {
-      super.next(nextState);
+    if (elfHooksRegistry.preStoreUpdate) {
+      nextState = elfHooksRegistry.preStoreUpdate(
+        currentState,
+        nextState,
+        this.name
+      );
     }
+
+    if (nextState !== currentState) {
+      this.state = nextState;
+
+      if (batchInProgress.getValue() && !this.batchInProgress) {
+        this.batchInProgress = true;
+
+        batchDone$.subscribe(() => {
+          super.next(this.state);
+          this.batchInProgress = false;
+        });
+      } else {
+        super.next(this.state);
+      }
+    }
+  }
+
+  getValue(): State {
+    return this.state;
   }
 
   reset() {