+++ packages/entities/CHANGELOG.md
@@ -2,6 +2,66 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [4.2.0](https://github.com/ngneat/elf/compare/entities-4.1.1...entities-4.2.0) (2022-03-17)
+
+
+### Features
+
+* **entities:** 🔥 getEntitiesCount, getEntitiesCountByPredicate ([288139e](https://github.com/ngneat/elf/commit/288139e34fd7e4ba21a5ba7bb08be05155920833)), closes [#194](https://github.com/ngneat/elf/issues/194)
+
+
+
+## [4.1.1](https://github.com/ngneat/elf/compare/entities-4.1.0...entities-4.1.1) (2022-03-17)
+
+
+### Bug Fixes
+
+* **entities:** 🐞 getAllEntitiesApply mapped return type ([da5ce29](https://github.com/ngneat/elf/commit/da5ce29b548a3d1ce9cc3dc95449acf317d62d6f))
+
+
+
+# [4.1.0](https://github.com/ngneat/elf/compare/entities-4.0.1...entities-4.1.0) (2022-03-12)
+
+
+### Features
+
+* **entities:** add getActiveEntity and getActiveEntities ([#192](https://github.com/ngneat/elf/issues/192)) ([9e4c16e](https://github.com/ngneat/elf/commit/9e4c16eb4ebd8e1917a240418b0069b828f85539))
+
+
+
+## [4.0.1](https://github.com/ngneat/elf/compare/entities-4.0.0...entities-4.0.1) (2022-03-09)
+
+
+### Bug Fixes
+
+* **entities:** 🐞 export getAllEntitiesApply ([5356f00](https://github.com/ngneat/elf/commit/5356f0070f81ec3a45880d2218c520dfe27010f3))
+
+
+
+# [4.0.0](https://github.com/ngneat/elf/compare/entities-3.4.0...entities-4.0.0) (2022-03-09)
+
+
+### Features
+
+* **entities:** 🔥 change getEntities to getAllEntities ([f20cb19](https://github.com/ngneat/elf/commit/f20cb19f1a27df4e73feb4ce4e339915956c9a26))
+* **entities:** 🔥 change selectAll to selectAllEntities ([e30b6d1](https://github.com/ngneat/elf/commit/e30b6d19ccac7759b9587edfff01c8c29248a738))
+* **entities:** 🔥 change selectAllApply to selectAllEntitiesApp ([b5a6566](https://github.com/ngneat/elf/commit/b5a6566b896ae2bb6e32e26fdae8fac7f1af12ce))
+
+
+### BREAKING CHANGES
+
+* **entities:** 🧨 change getEntities to getAllEntities
+* **entities:** 🧨 change selectAllApply to selectAllEntitiesApp
+* **entities:** 🧨 change selectAll to selectAllEntities
+
+
+
+# [3.4.0](https://github.com/ngneat/elf/compare/entities-3.3.0...entities-3.4.0) (2022-03-09)
+
+
+### Features
+
+* **entities:** 🔥 add getAllEntitiesApply query ([368871b](https://github.com/ngneat/elf/commit/368871bd52139e0c59d4ff6ed789282264d68bc9))
 # [3.3.0](https://github.com/ngneat/elf/compare/entities-3.2.1...entities-3.3.0) (2022-02-23)